changed to strings

+++ Cars/Car.cs
@@ -11,9 +11,9 @@ namespace Cars
         public string Name { get; set; }
         public double Displacement { get; set; }
         public int Cylinders { get; set; }
-        public double City { get; set; }
-        public double Highway { get; set; }
-        public double Combined { get; set; }
+        public int City { get; set; }
+        public int Highway { get; set; }
+        public int Combined { get; set; }
 
     }
 }
